feat: logging, refactor and bug fix (#14)
* feat: logging, refactor and bug fix

* feat: logging, refactor and bug fix - uom

* feat: logging, refactor and bug fix - workerType

* feat: logging, refactor and bug fix - project scope & status

* feat: logging, refactor and bug fix - material vendor, type & manufacturer

+++ src/main/java/in/oxidane/work/done/building/controller/UnitTypeController.java
@@ -60,7 +60,7 @@ public interface UnitTypeController {
     @GetMapping("/{id}")
     ResponseEntity<UnitTypeResponse> getUnitTypeById(
         @Parameter(description = "Unit type ID", required = true)
-        @PathVariable("id") Integer id
+        @PathVariable("id") Long id
     );
 
     /**
@@ -98,7 +98,7 @@ public interface UnitTypeController {
     @PutMapping("/{id}")
     ResponseEntity<UnitTypeResponse> updateUnitType(
         @Parameter(description = "Unit type ID", required = true)
-        @PathVariable("id") Integer id,
+        @PathVariable("id") Long id,
         @Parameter(description = "Updated unit type details", required = true)
         @RequestBody UnitTypeRequest request
     );
@@ -120,6 +120,6 @@ public interface UnitTypeController {
     @DeleteMapping("/{id}")
     ResponseEntity<Void> deleteUnitType(
         @Parameter(description = "Unit type ID", required = true)
-        @PathVariable("id") Integer id
+        @PathVariable("id") Long id
     );
-} 
+}
